Moved the application of join_road_weather function to the osm_query.R script
Was originally at the bottom of the Join_Road_Weather.R script. Made this change so that other scripts can make use of the join_road_weather.R script in slightly different ways (changing the naming of the output files). This will pave the way to allow us to pull in a broader set of crash observations for training, from multiple years.